Adjust the actions.yml-file to the repository (#1)
Adjust the actions.yml-file to the repository. Also used the main branch of CHP_modelling as this has now been updated (and updated values in tests correspondingly). An issue with the BioCHP constructor that was due to non-ordered Dicts was resolve by first extracting all resources to ordered vectors and then extracting related data onto this ordering.